// frontend/src/generated/model/houseChangeInterface.ts
/**
 * backend
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ActionInterface } from './actionInterface';


export interface HouseChangeInterface { 
    id: string;
    category: number;
    time: number;
    price: number;
    weather: boolean;
    action: ActionInterface;
}
export namespace HouseChangeInterface {
}