// frontend/src/generated/model/streetChangeInterface.ts
/**
 * backend
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ActionInterface } from './actionInterface';


export interface StreetChangeInterface { 
    id: string;
    length: number;
    action: ActionInterface;
}
export namespace StreetChangeInterface {
}